Fixed bug relating to quality scores/bases
Prior to this commit the VariantVisitor used the "AlignedBases" attribute
to get base sequences, and the PositionInAlignment field to extract
the sequence being studied. THIS IS WRONG. In fact, the PositionInAlignment
field is indexed against the QueryBases field (which is not padded for indels).

This commit changes this, and should make bases and quality scores line up.